+++ LibOS/shim/src/sys/shim_msgget.c
@@ -691,8 +691,8 @@ static int __store_msg_persist(struct shim_msg_handle* msgq) {
         goto out;
     }
 
-    int expected_size = sizeof(struct msg_handle_backup) + sizeof(struct msg_backup) * msgq->nmsgs +
-                        msgq->currentsize;
+    size_t expected_size = sizeof(struct msg_handle_backup)
+                           + sizeof(struct msg_backup) * msgq->nmsgs + msgq->currentsize;
 
     if (DkStreamSetLength(file, expected_size))
         goto err_file;
